fix: fix gold desynch in shared experience mode

+++ src/common/Messages/GameNetworkMessages/GoldChanged.cs
@@ -5,7 +5,7 @@ namespace MegabonkTogether.Common.Messages.GameNetworkMessages
     [MemoryPackable]
     public partial class GoldChanged : IGameNetworkMessage
     {
-        public float Amount { get; set; }
+        public int Amount { get; set; }
         public uint OwnerId { get; set; }
     }
 }
